fix user ownership, fix notification, and fix some bugs
User ownership was not being properly assigned based on the pipeline
user.
Notifications were only not going out if the user was in the pipeline
index, when it should also encompass all subpaths.
The pipeline user was not being created properly due to a typo.
Documents were failing to build in the menu due to a typo.