Set up url and views for approval pipeline
Still need templates and other structure, as well as proper actions.
Still not sure exactly what all admin actions will be or how exactly
they will work, but it's going to have to be at the very least the
ability to add arbitrary objects to a step and to remove arbitrary
objects from a step (and therefore the entire pipeline).