use release_ownership on cancel, return 410 on non-pending tickets
Cancel was simply changing the status to "cancel", but not actually
releasing the ticket permissions.  This would leave it entirely up to
the admin user to go and clear those out.  Now the ticket is correctly
released using the release_ownership method.  This does present a
potential leak, if the object is private in its step and is then
cancelled while in a "Published" state, but it's a rare enough scenario
that the admin should be unpublishing beforehand anyway, because it's
broken.
Tickets in a non-Pending state still allowed a state change through an
"Approve", "Reject", or "Cancel" action, which makes no sense.  This now
returns a 410 Gone status response.